+++ changelog.md
@@ -17,6 +17,8 @@ All notable changes to this project will be documented in this file.
 
 ## Work in progress
 
+## Version 2.0.0 - 2022/08/01
+
 ### Added
 
 - Overload `Int` comparisons with `StrictlyPositiveInt`, `NegativeInt` and